+++ index.html
@@ -1,6 +1,5 @@
 <!DOCTYPE html>
 <html lang="ru">
-<script src="C:\Users\~~~\1188157-kekstagram-4\js\functions.js"></script>
 <head>
   <meta charset="utf-8">
   <meta name="viewport" content="width=device-width,initial-scale=1">
@@ -232,6 +231,7 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-
+  <script src="C:\Users\~~~\1188157-kekstagram-4\js\functions.js"></script>
+  <script src="/js/main.js" type="module"></script>
 </body>
 </html>
